Fixed repository initialization bug.

+++ backend/src/main/java/com/portfolio/backend/user/User.java
@@ -11,22 +11,19 @@ import javax.persistence.Id;
 public class User {
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private Integer id;
-
+    private Long id;
     private String email;
 
     @JsonIgnore
     private String password;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
-
-
 
     public String getEmail() {
         return email;
